Clean up image aspect ratio
Set the parameters in the layout to scale the cover image without
messing up the aspect ratio. This should be revisited to make the
height of the tile scale properly to the aspect ratio we want to get.
Currently, we're just setting an arbitrary size and scaling the image,
cropping off any overflow. We really need to be more inteligent.