Covers can handle any # of digits
I changed the layout of the cover panels to be much more efficient. the
watchedCount and flavour text string are now smaller. They are also
baseline aligned, so it will align properly now without guessing. The
biggest change however is that both of the aforementioned views are now
nestled inside of a RelativeLayout, which allows the # of digits in the
watchedCount to freely scale to as many as needed, and the text will
always be propely positioned next to it. This change also allowed me to
clean up the cover adapter code, which no longer has to mess with the
margins, as moving the text is handled automatically by alightLeftOf.

..did I do it right?