Generic CoverAdapter
I updated CoverAdapter to cast to GenericMALRecord, as well as adding an
abstract method getAmountConsumed() to GenericMALRecord, which is to be
implemented by AnimeRecord and MangaRecord to return a "consumed"
amount, in the case of AnimeRecord, the watched count. AnimeRecord's
implementation is already handled, and everything should continue to
build and run fine without errors.