Merged Cover Actions code
I merged in the Cover Actions code from the working branch so that manga
support can hopefully avoid any conflicts, since the Cover Actions did
require me to manually merge once because there was some modifications
in some of the files that are also used in manga support (ie.
CoverAdapter). Since Cover Actions are done, merging this in felt like a
smart idea in order to prevent conflicts. Also I updated the .gitignore
with the .settings folder for my Eclipse.