Added "Mark Completed" cover action.
I added the mark completed cover action, as well as moved some of the
final strings from Anime and MangaRecord into the GenericRecord. Also
made the getTotal() method part of GenericMALRecord, so I could access
it from cover adapter to automatically mark as completed if +1 Watched
would finish the series.