Scrolling Optimizations
I rewrote CoverAdapter to use the ViewHolder pattern, greatly reducing
the number of called to findViewById(). That change alone made a pretty
good performance improvement, but I didn't stop there. I went into
ImageDownloader and put decodeFile() onto a separate thread, as it was
running on the UI thread and was the sole reason the scrolling was
choppy. Now the list scrolls smooth as butter. TODO: A loading image
that's not a black panel.