App Credits
Since making a new layout for an about screen is too much effort at the
moment, for now I shuffled some things around in the settings. The
settings screen now mentions the app under the credits as a open source
project. Tapping it opens a new screen with contributors to the project.