Fix Login Crash Bug
Finally figured out what was causing a few users to crash, the culprit
being accidentally inserting a trailing space after their username while
signing in, resulting in an IllegalCharacterException when creating the
URI to sign in with. Simple fix, just silently delete the spaces when
they sign in. This one's been haunting me since all the way back to when
I released the very first version (pre-rewrite), so it feels good to
finally fix this.