ImageDownloader Crash Fix
I had multiple reports of a crash due to ImageDownloader, and the cause
appears to have been a NPE when calling cover. I can't reproduce the
crash, but as cover is a weak reference, the likely cause was it was
garbage collected. Added a check to only run the code if the reference
isn't null, hopefully that should fix it.