Fix Crash on start
During commit of 149efb77ac, res/values-large/styles.xml got added. This
was being used by some devices and causing a non-ActionBar theme to be
used, which caused a crash when trying to access the action bar in the
application. This change removes the bad file and adjusts
res/values/styles.xml to use the Holo style like it should.