ABS Related Changes
The project now uses ActionBarSherlock classes in order to support all
the way back to Android 2.2. This commit sees the beginning of the
conversion with First Run, Home, and ItemGridFragment converted.
There's also a nasty bug where you can't log in yet because of some
crazy error.

Also, because ABS is a separate Android project acting a library, to use
this repo now you have to get it working in your local IDE (or compiler)
manually.