Added Scores to the Info Card
"My Score" and "MAL Score" and now part of the info card. As part of
this, since the MAL Score was never downloaded before I updated the
relevant methods to fetch the information if it doesn't exist on the
device. I also slightly modified the getMyScore() method in
GenericMALRecord to return "Not Rated Yet" if the entry is 0 - not
rated.