Auto-adjust grid height
Now, we calculate and auto-adjust the height of the grid cells to match
our desired 0.7 aspect ratio. This should allow most cover images to
properly scale without much cropping being done. Some longer or wider
images will still be cropped, but there isn't much that can be done
about that.

Note that this is fairly rough. It can probably be refined a bit more.