Card Desgin Tweaks (and a bugfix)
The "Info" and "Progress" cards are now "MAL Stats" and "Personal
Stats". Watch Status and My Score have also moved from MAL Stats to
Personal Stats.

Bugfix: By displaying "Not Yet Available" through the getMyScore()
method, I unintentionally was storing that in the database
because of how we save items. Reverted getMyScore() to just returning
the score and created a getFormattedScore() method for handling what it
was previously.