Fix NPE Error in Score
In some cases, the application won't have the item score. Passing a null
to the view results in a NullPointerException. This little kludge causes
a zero to be returned if we don't have the score, preventing the
exception and avoiding an application crash.

This should probably be handled better later, perhaps by fetching the
score if we don't have it.