Change Scores from Strings to Numbers
The score values should be treated as numbers, not strings. Better typing
should help prevent exception errors. Now, the MAL Member Score is a
float and the user's score is an integer.

Also, since there is now a dedicated "card" for scores, this commit also
removes the display of the member score from the general info card.