Add Rating Dialog (1/2)
Like previous commits like this, this adds the first part
to the rating dialog, that is the dialog itself. Everything
works, and it does indeed pass a value back to DetailView,
but DetailView doesn't do anything with it.

Additionally, I spent way too long making this work, but I
extended RatingBar to make the flavour text update as you
slide your finger across it. Looks super cool and neat.