Add Remove from List (1/2)
This first commit for this feature adds all of UI and
workflow to make this work. The remove action will never
be shown in the action bar, and thus must be invoked on
purpose. In addition, a confirmation dialog will appear
asking the user to confirm their choice, as the action is
destructive. Currently non-functional, 2/2 will make it work.