Fixed Status String Baseline In Absence of Count
On grid covers, the status string was aligned to the baseline of the
item count. This caused a different alignment for statuses where the
count was hidden. It was especially noticible for views which mixed
statuses. Adding padding to the bottom of the status string now keeps
alignment the same.