Restore compatibility with JRE1.6 (Java 6)
The commits that added search used methods only available
in JRE1.7 (Java 7), this commit fixes them to be JRE1.6 compliant.

In fact, I'm not even sure how Apkawa got it to compile, as
it refused to compile on my end, and Android shouldn't even be
able to run JRE1.7 code, as Dalvik is based on JRE1.6.

TODO is a fix a method introduces in MALManager that uses an API only
available at level 11 or higher. I've added a version check for the
call, and currently returning null on version prior to 11. While this
is breaking functionality for pre-3.0 devices, it does allow the
code to compile. A suitable replacement should be added in the method
to ensure Froyo and Gingerbread users can use the app.