Add to list UI
Added the UI flow for adding an item to list. Basically,
we check if the record's watch status is null, which is
only a possibility if it isn't on a list, then set the
action bar to contain an add button, and hide extraneous
elements (like score, progress) as they don't matter
when it isn't on a list.

I've also added a method in this commit to revert the UI
changes after a record is added, but this is obviously not
functioning at the moment due to the add button not doing
anything :)