Add to list!
This was a long time coming, but as of this commit, after
searching for an anime or manga, you can simply hit the +
to add it to your list. The current behaviour defaults
to "watching" or "reading", though I may make this
customizable. Or if we can learn start date, have it
"planned" before the start date and "watching" after.
I dunno, just pitching ideas.