"Everything" sort fix
As records that were searched for and loading in a
detail view are stored in the database, the everything list
includes these entries that were not added. As a fix, I've
updated the database searching to search for any of
the statuses that something on a list can be, therefore
they no longer appear.