Fix NPE in Old Android
As best as I can figure out, there is some kind of threading issue in
Android 3.0 and below where the typeMap somehow gets used for both
manga and anime records, making values not match properly. This leads
to null values in areas that shouldn't have them. This is fixed by
always building the typeMap when getTypeMap is called.

Fixes #51