Beautiful App Launching + Optimizations
This commit serves to make the app launch as seamless
as possible. Previously, the background that was drawn
when the app launched wasn't drawn immediately because it
was part of the layout, not the window. This commit moves
the background to be part of the app's style, thus no more
jarring white -> gray transition.

As a result, I've also eliminated an entire layer of
overdraw throughout the app, so devices with lower-end
GPUs should have an easier time rendering the layouts.