Connectivity State Croutons
This commit adds croutons that appear in the main UI
to alert the user that we will be going into or leaving
offline mode. The checks happen onResume() and also
on the "CONNECTIVITY_CHANGE" broadcast in order to cover
all possible scenarios when connectivity loss could arise.

In addition, this commit also disables the option to sync
while offline by checking a flag that is set by the
connectivity check. Todo is make everything work offline.