Croutons queuing outside the app fix
Basically, if network connectivity changed outside the
app, the BroadcastReceiver was still registered and
was still listening for events, resulting in a potentially
long queue of croutons when the app was opened.

This commit changes the behaviour such that the receiver
is only listening while in the Home activity, and
unregistered on leaving. This completes the desired
behaviour of only seeing the Croutons when a state change
occurs, or when entering the Home screen via the check
onResume().