Covers: Add Fade In Animation
Low Priority == I felt like doing it right now.

This commit adds a lovely fade in animation to the covers
whenever we need to load them in, be it from the network
or decoding them from file. This also contributes to how
nice the app looks on launch. There is no longer any
jarring elements to induce eye-cancer on startup :)