Bring our own Support Library
Out of the blue, Eclipse has deciced to have a seizure
and realize that Crouton is actually an issue when it
rightfully should be. The reason for this error is because
we are providing an extremely old version of the support
library from ABS. Crouton was using a constant that had
yet to exist, so now we have to bring our own library.

Also updated BUILDING to reflect the new steps.