Search Tweaks and Fixes
Prevented crashes if the API has issues on searching.

Also changed search to now search both anime AND manga at the
same time. In addition, if there is no results for either a
crouton is displayed. If there is only results for one, the
pager autoselects the one with results and slides to it.

Also, keyboard is hidden after subsequent searches within the
SearchActivity, and the SearchView is collapsed upon returning
to the main activity. This is optimal behaviour and will make
people (me) happy.