Always Pass Context to getRecords
Context isn't passed to our background tasks in Android versions below 4. Change the calls to getRecords to explicitly always pass context to the method and then pass context to the methods it calls. We can then safely create MALManager if it's null (Android below 4) without an NPE when on our first login.

Fixes #54