Consistent Icons Prep
We can use Android's resource system to handle #57, so I created
some -v11 folders and moved the only resource (Refresh) that I know
could be in either the action bar or the 2.3 menu depending on
available space to it. Should just need to add 2.3 style resources
and we'll be set.