Fix for loading indicator after first init screen
After the first init there are two Tasks running: one GETLIST (always
called because of displaying the Home-screen) and one FORCESYNC (called
because it is the first sync). This led to the problem that the loading
indicator was hidden to fast because the GETLIST task returned faster
than the other one (because the database is emtpy) and hid the loading
indicator.
Now the loading indicator does not get hidden when there is also a
forced sync that has not finished.