Gradle configuration update for Android Studio >= 0.6
This makes the gradle configuration compatible to Android Studio 0.6.
Also every dependency is loaded from maven repos now, SherlockNavDrawer
was the last one that was not fetched from maven but I noticed that the
author has a maven repo for this so I updated the configuration.
This was necessary because the gradle config in the SherlockNavDrawer
submodule was not compatible to Android Studio >= 0.6.