Unify Anime/MangaNetworkTask callback interface
The callback functions always contain the same logic for Anime and
Manga, so they can be unified in on function. This removes some
duplicated code and avoids repetitions if something needs to be changed
in the callbacks.
So now there is only one callback for Anime/MangaNetworkTask:
onNetworkTaskFinished, without type separation. To determine the list
type the function has the additional ListType parameter "type".