Gradle Fixes
Some small fixes to allow Gradle to work. Google released v20 and 21 of
the support library, and as we were not forcing a specific version, it
was choosing the newest. This caused issues with Crouton and also SDK
versioning. Force support library 19 to avoid the error. Additionally,
update the version of Crouton to pull in some bug fixes from it.