Fix possible SQLite error 1802
After deleting the database (on logout), no new one was created because
the MALSqlHelper was not recreated for two reasons:
- the MALSqlHelper is designed as singleton (object is retrieved by
  calling getHelper()), but instance was not unset after deleting the
  database
- malSqlHelper was defined as static in DatabaseManager, so even fixing
  the first point would not fix the bug, because all instances of
  DatabaseManager share the same malSqlHelper

Now MALSqlHelper has the new function deleteDatabase() that deletes the
database and sets instance to null so the next call to getHelper() will
recreate the database. Also malSqlHelper is not defined as static in
DatabaseManager. Also cancel running NetworkTasks on logout to make sure
that the tasks are not accessing the database when it is deleted.