Save Anime-/Mangalist in relation table
This allows us to store the lists of friends too! Updated all related
functions (they now have a username or userId parameter to define the
user).
Also removed the recordId field from the anime and manga table: as the
ID's provided by the API are unique we can use them directly.