Fix for setDirty(true) being always called
Because fromUser was not checked in the onRatingChanged()-callback the
code was always executed, not only if the rating was changed by the
user. This means the record was always flagged as dirty (which caused
unneccessary API requests), because this callback is also called if the
value is changed programmatically.